move test to avoid gopkg.in dependencies
